refactor: 에러 파일 공통 패키지로 이동
