+++ src/main/java/com/devnity/devnity/common/error/exception/EntityNotFoundException.java
@@ -1,4 +1,4 @@
-package com.devnity.devnity.web.error.exception;
+package com.devnity.devnity.common.error.exception;
 
 public class EntityNotFoundException extends BusinessException {
