+++ src/main/java/com/devnity/devnity/common/error/exception/S3UploadException.java
@@ -1,4 +1,4 @@
-package com.devnity.devnity.web.error.exception;
+package com.devnity.devnity.common.error.exception;
 
 public class S3UploadException extends BusinessException  {
   public S3UploadException() {
