+++ src/main/java/com/devnity/devnity/domain/admin/service/AdminGenerationService.java
@@ -1,7 +1,7 @@
 package com.devnity.devnity.domain.admin.service;
 
-import com.devnity.devnity.web.error.exception.EntityNotFoundException;
-import com.devnity.devnity.web.error.exception.ErrorCode;
+import com.devnity.devnity.common.error.exception.EntityNotFoundException;
+import com.devnity.devnity.common.error.exception.ErrorCode;
 import com.devnity.devnity.domain.admin.dto.request.GenerationRequest;
 import com.devnity.devnity.domain.admin.dto.response.GenerationResponse;
 import com.devnity.devnity.domain.user.entity.Generation;
