+++ src/main/java/com/devnity/devnity/domain/admin/service/AdminInvitationService.java
@@ -1,7 +1,7 @@
 package com.devnity.devnity.domain.admin.service;
 
-import com.devnity.devnity.web.error.exception.ErrorCode;
-import com.devnity.devnity.web.error.exception.InvalidValueException;
+import com.devnity.devnity.common.error.exception.ErrorCode;
+import com.devnity.devnity.common.error.exception.InvalidValueException;
 import com.devnity.devnity.domain.admin.dto.InvitationDto;
 import com.devnity.devnity.domain.admin.dto.request.InvitationRequest;
 import com.devnity.devnity.domain.admin.entity.Invitation;
