+++ src/main/java/com/devnity/devnity/web/user/service/UserRetrieveService.java
@@ -1,9 +1,9 @@
 package com.devnity.devnity.web.user.service;
 
-import static com.devnity.devnity.web.error.exception.ErrorCode.ENTITY_NOT_FOUND;
-import static com.devnity.devnity.web.error.exception.ErrorCode.USER_NOT_FOUND;
+import static com.devnity.devnity.common.error.exception.ErrorCode.ENTITY_NOT_FOUND;
+import static com.devnity.devnity.common.error.exception.ErrorCode.USER_NOT_FOUND;
 
-import com.devnity.devnity.web.error.exception.EntityNotFoundException;
+import com.devnity.devnity.common.error.exception.EntityNotFoundException;
 import com.devnity.devnity.domain.introduction.dto.IntroductionDto;
 import com.devnity.devnity.domain.introduction.entity.Introduction;
 import com.devnity.devnity.domain.mapgakco.service.MapService;
